requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
packaging==23.0
pkgutil-resolve-name==1.3.10
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
swagger-spec-validator==3.0.3
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
zipp==3.15.0
